+++ boot/include/multiboot.h
@@ -346,29 +346,29 @@ struct Multiboot_Info_BootDevice {
     uint32_t reserved;
 }__attribute__((packed));
 
-extern bool      Multiboot_CheckForValidMBI     (uintptr_t mbi_addr);
-extern uintptr_t Multiboot_FindMBITagAddress    (uintptr_t mbi_addr, uint32_t tag_type);
-extern bool      Multiboot_CreateEmptyMBI       (uintptr_t mbi_addr);
-extern bool      Multiboot_TerminateTag         (uintptr_t mbi_addr, uintptr_t tag_addr);
-extern uintptr_t Multiboot_GetHeader            (uintptr_t start_addr, size_t size);
+extern bool     Multiboot_CheckForValidMBI     (uint32_t mbi_addr);
+extern uint32_t Multiboot_FindMBITagAddress    (uint32_t mbi_addr,   uint32_t tag_type);
+extern bool     Multiboot_CreateEmptyMBI       (uint32_t mbi_addr);
+extern bool     Multiboot_TerminateTag         (uint32_t mbi_addr,   uint32_t tag_addr);
+extern uint32_t Multiboot_GetHeader            (uint32_t start_addr, uint32_t size);
 
-extern bool      Multiboot_LoadKernelFile       (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
-extern bool      Multiboot_LoadKernel           (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
-extern bool      Multiboot_LoadModules          (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_LoadKernelFile       (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_LoadKernel           (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_LoadModules          (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
 
-extern bool      Multiboot_SaveBootLoaderInfo   (uintptr_t mbi_addr);
-extern bool      Multiboot_SaveBootCommand      (uintptr_t mbi_addr);
-extern bool      Multiboot_SaveMemoryMaps       (uintptr_t mbi_addr);
-extern bool      Multiboot_SaveBasicMemoryInfo  (uintptr_t mbi_addr);
-extern bool      Multiboot_SaveAPMInfo          (uintptr_t mbi_addr);
-extern bool      Multiboot_SaveSMBIOSInfo       (uintptr_t mbi_addr);
-extern bool      Multiboot_SaveACPIInfo         (uintptr_t mbi_addr, bool old);
-extern bool      Multiboot_SaveELFSectionHeaders(uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
-extern bool      Multiboot_SaveBootDeviceInfo   (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
-extern bool      Multiboot_SaveGraphicsInfo     (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
-extern bool      Multiboot_SaveLoadBaseAddress  (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
-extern bool      Multiboot_SaveInfo             (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_SaveBootLoaderInfo   (uint32_t mbi_addr);
+extern bool     Multiboot_SaveBootCommand      (uint32_t mbi_addr);
+extern bool     Multiboot_SaveMemoryMaps       (uint32_t mbi_addr);
+extern bool     Multiboot_SaveBasicMemoryInfo  (uint32_t mbi_addr);
+extern bool     Multiboot_SaveAPMInfo          (uint32_t mbi_addr);
+extern bool     Multiboot_SaveSMBIOSInfo       (uint32_t mbi_addr);
+extern bool     Multiboot_SaveACPIInfo         (uint32_t mbi_addr, bool old);
+extern bool     Multiboot_SaveELFSectionHeaders(uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_SaveBootDeviceInfo   (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_SaveGraphicsInfo     (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_SaveLoadBaseAddress  (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_SaveInfo             (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
 
-extern bool      Multiboot_Boot                 (uintptr_t mbi_addr, struct Boot_KernelInfo* kernel_info);
+extern bool     Multiboot_Boot                 (uint32_t mbi_addr, struct Boot_KernelInfo* kernel_info);
 
 #endif
